requirements.txt
# CodeArchitect AI Dependencies

# Core
python-dotenv==1.0.0
pyyaml==6.0.1

# OpenAI (optional - only needed if using provider: openai)
openai>=1.12.0

# Local Embeddings (optional - only needed if using provider: local)
sentence-transformers>=2.3.1
torch>=2.0.0

# Utilities
tiktoken>=0.5.2
tenacity==8.2.3

# Development
pytest>=8.0.0
black>=24.1.1
flake8>=7.0.0

